Use standard ints

+++ lib/PickleServo/PickleServo.cpp
@@ -8,18 +8,18 @@ void PickleServo::attach(uint8_t pin, uint8_t start)
   this->write(start);
 }
 
-void PickleServo::move(uint8_t angle, uint32_t time)
+void PickleServo::move(int32_t angle, uint32_t time)
 {
   // Calculate whether the angle should increment or decrement
-  int start = this->read();
-  int increment = (start < angle) ? 1 : -1;
+  int32_t start = this->read();
+  int32_t increment = (start < angle) ? 1 : -1;
 
   // Calculate the delay time in microseconds (milliseconds is not precise enough)
-  int delayTime = 1000 * time / abs(angle - start);
+  uint32_t delayTime = 1000 * time / abs(angle - start);
 
   if (delayTime)
   {
-    for (int current = start; current != angle; current += increment)
+    for (int32_t current = start; current != angle; current += increment)
     {
       this->write(current);
       delayMicroseconds(delayTime);
